ADDED: NewSBuild.props` and `Dgmjr.MSBuild.Extensions.targets` were added to the project.
FIXED The `` property in thejr.MSBuild.Extensions.cs remove `net6.0` and ` the `Dgmjr.MSBuild.Extensions.csproj` file was updated the `using System.Collections.Generic;` namespace.

UPDATED: The `Extensions.cs` file was modified to include package references and change the `TargetFramework` to `netstandard2.0`.

UPDATED: The `GetCommandLineArgs.props` file was modified to include the `UsingTask` element for the `GetCommandLineArgs` task.

UPDATED: The `GetCommandLineArgs.targets` file was modified to define the `GetCommandLineArgs` target.

UPDATED: The `GetCommandLineArgs.xml` file was modified to update the summary for the `ThisAssembly.Git.Commit` member and add new members related to git information.

UPDATED: The `CommonPropertyNames.cs` file was modified to remove unnecessary comment lines.